NXP-22217: separate pure blob management interface
