+++ nuxeo-core/nuxeo-core-test/src/test/resources/OSGI-INF/test-blob-dispatcher.xml
@@ -1,6 +1,15 @@
 <?xml version="1.0"?>
 <component name="org.nuxeo.ecm.core.blob.test-blob-dispatcher" version="1.0.0">
   <extension target="org.nuxeo.ecm.core.blob.BlobManager" point="configuration">
+    <blobprovider name="test2">
+      <class>org.nuxeo.ecm.core.blob.binary.DefaultBinaryManager</class>
+      <property name="path">binaries2</property>
+    </blobprovider>
+    <blobprovider name="dummy">
+      <class>org.nuxeo.ecm.core.DummyBlobProvider</class>
+    </blobprovider>
+  </extension>
+  <extension target="org.nuxeo.ecm.core.blob.DocumentBlobManager" point="configuration">
     <blobdispatcher>
       <class>org.nuxeo.ecm.core.blob.DefaultBlobDispatcher</class>
       <property name="dc:format=video">test2</property>
@@ -10,12 +19,5 @@
       <property name="ecm:repositoryName=test,ecm:lifeCycleState=approved">test2</property>
       <property name="default">test</property>
     </blobdispatcher>
-    <blobprovider name="test2">
-      <class>org.nuxeo.ecm.core.blob.binary.DefaultBinaryManager</class>
-      <property name="path">binaries2</property>
-    </blobprovider>
-    <blobprovider name="dummy">
-      <class>org.nuxeo.ecm.core.DummyBlobProvider</class>
-    </blobprovider>
   </extension>
 </component>
