+++ nuxeo-core/nuxeo-core/src/main/resources/META-INF/MANIFEST.MF
@@ -39,6 +39,7 @@ Nuxeo-Component: OSGI-INF/CoreService.xml,
   OSGI-INF/character-filtering-service.xml,
   OSGI-INF/character-filtering-service-contrib.xml,
   OSGI-INF/blobmanager-service.xml,
+  OSGI-INF/documentblobmanager-service.xml,
   OSGI-INF/uidgenerator-service.xml,
   OSGI-INF/CoreSessionService.xml,
   OSGI-INF/scheduler-contrib.xml
