+++ nuxeo-core/nuxeo-core/src/test/java/org/nuxeo/ecm/core/blob/DummyBlobProvider.java
@@ -30,7 +30,6 @@ import java.util.concurrent.atomic.AtomicLong;
 import org.apache.commons.io.IOUtils;
 import org.nuxeo.ecm.core.api.Blob;
 import org.nuxeo.ecm.core.blob.BlobManager.BlobInfo;
-import org.nuxeo.ecm.core.model.Document;
 
 /**
  * Dummy storage in memory.
@@ -79,7 +78,7 @@ public class DummyBlobProvider extends AbstractBlobProvider {
     }
 
     @Override
-    public String writeBlob(Blob blob, Document doc) throws IOException {
+    public String writeBlob(Blob blob) throws IOException {
         byte[] bytes;
         try (InputStream in = blob.getStream()) {
             bytes = IOUtils.toByteArray(in);
